Made a config and edit how it works
Made it so the user can just put in a discord hook into the form it self and send it.

It got some stuff the user would need to edit to make sure its to what they like. That is very easy done in the config section (bot name did not want to be there. You can find it bit info the function)